move statistics logic
